refactor: split schema parser helpers by responsibility
